refactor(runtime): split resolution.rs into cohesive modules (§7-8)
Split the 2804-line runtime/resolution.rs (over the 500-line limit) into
eight cohesive modules under src/runtime/. All but one are under 500 lines;
resolution_call_sub.rs (651) holds the single ~578-line call_sub_value
method, which cannot be split at a function boundary. All methods are impl
Interpreter, so the split is visibility-only.

- resolution.rs (239)              - function/token-def resolution, MRO helpers
- resolution_call_sub.rs (651)     - call_sub_value + phaser-split + promise/supply
- resolution_method.rs (446)       - method resolution + type-distance ranking
- resolution_map_grep_rw.rs (423)  - eval_map/grep_over_items_rw (mutated)
- resolution_eval.rs (363)         - eval/compile/run compiled block values
- resolution_map_grep.rs (348)     - eval_map/grep_over_items + FirstMatcher
- resolution_private_method.rs (187)- private-method resolution
- resolution_lazy.rs (169)         - force_lazy_list / force_lazy_io_lines

Pure relocation. Three private methods bumped to pub(crate); the FirstMatcher
trait module is pub(crate) (used by vm_native_first, whose import path was
updated). The fn multiset is byte-identical. No behavior change.
