+++ src/runtime/mod.rs
@@ -257,6 +257,13 @@ mod registration_class;
 pub(crate) mod registration_sub;
 mod registry;
 pub(crate) mod resolution;
+mod resolution_call_sub;
+mod resolution_eval;
+mod resolution_lazy;
+pub(crate) mod resolution_map_grep;
+mod resolution_map_grep_rw;
+mod resolution_method;
+mod resolution_private_method;
 mod run;
 mod run_dist;
 mod run_modules;
