+++ src/runtime/resolution.rs
@@ -1,27 +1,11 @@
 use super::*;
 use crate::symbol::Symbol;
 
-type CompiledFnMap = std::collections::HashMap<String, crate::opcode::CompiledFunction>;
-type ProtectBlockCompiled = std::sync::Arc<crate::opcode::CompiledCode>;
-type ProtectBlockCompiledFns = std::sync::Arc<CompiledFnMap>;
-type ProtectBlockCapturedBindings = std::sync::Arc<Vec<(usize, String)>>;
-type ProtectBlockWritebackBindings = std::sync::Arc<Vec<(usize, String)>>;
-type ProtectBlockCapturedNames = std::sync::Arc<Vec<String>>;
-/// (pre_phasers, enter_phasers, success_queue, failure_queue, post_phasers, body_main)
-pub(super) type SplitPhasers = (
-    Vec<Stmt>,
-    Vec<Stmt>,
-    Vec<Stmt>,
-    Vec<Stmt>,
-    Vec<Stmt>,
-    Vec<Stmt>,
-);
-
 impl Interpreter {
     pub(crate) const LAZY_GATHER_TAKE_LIMIT_SIGNAL: &str =
         "__mutsu_lazy_gather_take_limit_reached__";
 
-    fn is_stub_method_body(body: &[Stmt]) -> bool {
+    pub(crate) fn is_stub_method_body(body: &[Stmt]) -> bool {
         let filtered: Vec<_> = body
             .iter()
             .filter(|s| !matches!(s, Stmt::SetLine(_)))
@@ -252,2553 +236,4 @@ impl Interpreter {
             .proto_tokens
             .contains(&format!("GLOBAL::{}", name))
     }
-
-    pub(super) fn resolve_method(
-        &mut self,
-        class_name: &str,
-        method_name: &str,
-        arg_values: &[Value],
-    ) -> Option<MethodDef> {
-        self.resolve_method_with_owner(class_name, method_name, arg_values)
-            .map(|(_, def)| def)
-    }
-
-    fn method_args_match_for_invocant(
-        &mut self,
-        class_name: &str,
-        def: &MethodDef,
-        arg_values: &[Value],
-        role_bindings: Option<&HashMap<String, Value>>,
-        invocant: Option<&Value>,
-    ) -> bool {
-        let saved_env = self.env.clone();
-        if let Some(bindings) = role_bindings {
-            for (name, value) in bindings {
-                self.env.insert(name.clone(), value.clone());
-            }
-        }
-        for pd in &def.param_defs {
-            if !(pd.is_invocant || pd.traits.iter().any(|t| t == "invocant")) {
-                continue;
-            }
-            if let Some(constraint) = pd.type_constraint.as_deref() {
-                if let Some(captured_name) = constraint.strip_prefix("::") {
-                    self.bind_type_capture(
-                        captured_name,
-                        &Value::Package(Symbol::intern(class_name)),
-                    );
-                }
-                // Check type constraint on the invocant (including :U/:D smileys).
-                // Resolve ::?CLASS to the actual class name for matching.
-                // Pure type captures (e.g. ::T) don't constrain the invocant.
-                if let Some(inv) = invocant {
-                    let resolved = constraint.replace("::?CLASS", class_name);
-                    let is_type_capture = resolved.starts_with("::");
-                    if !is_type_capture && !self.type_matches_value(&resolved, inv) {
-                        self.env = saved_env;
-                        return false;
-                    }
-                }
-            }
-        }
-        let args_match = self.method_args_match(arg_values, &def.param_defs);
-        self.env = saved_env;
-        args_match
-    }
-
-    pub(crate) fn resolve_method_with_owner(
-        &mut self,
-        class_name: &str,
-        method_name: &str,
-        arg_values: &[Value],
-    ) -> Option<(String, MethodDef)> {
-        self.resolve_method_with_owner_impl(class_name, method_name, arg_values, None)
-    }
-
-    pub(crate) fn resolve_method_with_owner_invocant(
-        &mut self,
-        class_name: &str,
-        method_name: &str,
-        arg_values: &[Value],
-        invocant: &Value,
-    ) -> Option<(String, MethodDef)> {
-        self.resolve_method_with_owner_impl(class_name, method_name, arg_values, Some(invocant))
-    }
-
-    fn resolve_method_with_owner_impl(
-        &mut self,
-        class_name: &str,
-        method_name: &str,
-        arg_values: &[Value],
-        invocant: Option<&Value>,
-    ) -> Option<(String, MethodDef)> {
-        self.dispatch_ambiguous = false;
-        let role_bindings = self.registry().get_role_param_bindings(class_name);
-        let mro = self.class_mro(class_name);
-        // Collect all matching multi candidates across the MRO, then pick the
-        // most specific one by type hierarchy distance.
-        let mut all_matches: Vec<(String, MethodDef)> = Vec::new();
-        // Track whether a non-multi submethod was found on an ancestor
-        // (submethods block MRO search for their class but not for
-        // descendants).
-        let mut submethod_blocks = false;
-        for cn in &mro {
-            // Hoist the clone to a `let` so the registry read guard drops here,
-            // before method_args_match_for_invocant re-enters (&mut self). An
-            // `if let` scrutinee would otherwise keep the temporary guard alive
-            // across the whole block.
-            let overloads = self
-                .registry()
-                .get_method_overloads(cn.as_str(), method_name);
-            if let Some(overloads) = overloads {
-                let any_multi = overloads.iter().any(|d| d.is_multi);
-                let mut first_visible_non_multi: Option<MethodDef> = None;
-                // Check if all overloads are submethods on an ancestor class
-                let all_submethods = overloads.iter().all(|d| d.is_my);
-                let is_ancestor = cn != class_name;
-                for def in overloads {
-                    if def.is_private {
-                        continue;
-                    }
-                    // Submethods are NOT inherited: skip if defined on an
-                    // ancestor class rather than the receiver's own class.
-                    if def.is_my && is_ancestor {
-                        continue;
-                    }
-                    if !any_multi && first_visible_non_multi.is_none() {
-                        first_visible_non_multi = Some(def.clone());
-                    }
-                    let args_match = self.method_args_match_for_invocant(
-                        class_name,
-                        &def,
-                        arg_values,
-                        role_bindings.as_ref(),
-                        invocant,
-                    );
-                    if args_match {
-                        if !any_multi && all_matches.is_empty() {
-                            // Non-multi: return the first match immediately,
-                            // but only if no multi candidates were already
-                            // collected from a child class in the MRO.
-                            return Some((cn.clone(), def));
-                        }
-                        all_matches.push((cn.clone(), def));
-                    }
-                }
-                // A non-multi submethod on an ancestor blocks the MRO search
-                // (the method name exists but no candidate matched and
-                // submethods are not inherited).
-                if !any_multi && is_ancestor && all_submethods {
-                    submethod_blocks = true;
-                    continue;
-                }
-                // Method name is present on this class, but no candidate matched.
-                // For non-multi methods, stop here — a subclass override
-                // hides the parent's version.
-                if !any_multi && all_matches.is_empty() {
-                    return first_visible_non_multi.map(|def| (cn.clone(), def));
-                }
-            }
-        }
-        if all_matches.len() <= 1 {
-            return all_matches.into_iter().next();
-        }
-        // Pick the candidate with the smallest type hierarchy distance
-        let mut best_idx = 0;
-        let mut best_dist = self.method_candidate_type_distance(arg_values, &all_matches[0].1);
-        for (i, (_, def)) in all_matches.iter().enumerate().skip(1) {
-            let dist = self.method_candidate_type_distance(arg_values, def);
-            if dist < best_dist {
-                best_dist = dist;
-                best_idx = i;
-            }
-        }
-        // `is default` trait tie-breaking: among candidates with the same
-        // best distance, prefer the one marked `is default`.
-        let tied: Vec<usize> = all_matches
-            .iter()
-            .enumerate()
-            .filter(|(_, (_, def))| {
-                self.method_candidate_type_distance(arg_values, def) == best_dist
-            })
-            .map(|(i, _)| i)
-            .collect();
-        if tied.len() > 1 {
-            let best_where = tied
-                .iter()
-                .map(|&i| {
-                    all_matches[i]
-                        .1
-                        .param_defs
-                        .iter()
-                        .filter(|p| p.where_constraint.is_some())
-                        .count()
-                })
-                .max()
-                .unwrap_or(0);
-            let narrowed: Vec<usize> = tied
-                .iter()
-                .copied()
-                .filter(|&i| {
-                    all_matches[i]
-                        .1
-                        .param_defs
-                        .iter()
-                        .filter(|p| p.where_constraint.is_some())
-                        .count()
-                        == best_where
-                })
-                .collect();
-            if let Some(&i) = narrowed.first() {
-                best_idx = i;
-            }
-            // Invocant specificity tie-break: candidates whose argument-type
-            // distance is equal can still differ in how derived their owning
-            // class is. A `multi method` defined on the receiver's own class is
-            // more specific than one inherited from an ancestor, so the
-            // most-derived owner (earliest in the MRO) wins rather than being
-            // treated as ambiguous.
-            let mro_index = |owner: &str| -> usize {
-                mro.iter()
-                    .position(|cn| cn.as_str() == owner)
-                    .unwrap_or(usize::MAX)
-            };
-            let best_mro = narrowed
-                .iter()
-                .map(|&i| mro_index(all_matches[i].0.as_str()))
-                .min()
-                .unwrap_or(usize::MAX);
-            let mro_narrowed: Vec<usize> = narrowed
-                .iter()
-                .copied()
-                .filter(|&i| mro_index(all_matches[i].0.as_str()) == best_mro)
-                .collect();
-            if let Some(&i) = mro_narrowed.first() {
-                best_idx = i;
-            }
-            let mut default_winner = false;
-            for &i in &tied {
-                if all_matches[i].1.is_default {
-                    best_idx = i;
-                    default_winner = true;
-                    break;
-                }
-            }
-            // Ambiguous dispatch: two or more candidates remain equally specific
-            // (same type distance, same number of `where` constraints, and the
-            // same most-derived owner class) and none is marked `is default`.
-            // Raku raises X::Multi::Ambiguous here.
-            if !default_winner && mro_narrowed.len() > 1 {
-                self.dispatch_ambiguous = true;
-            }
-        }
-        let _ = submethod_blocks; // used for control flow above
-        Some(all_matches.remove(best_idx))
-    }
-
-    /// Compute the type distance of a method's param constraints from the
-    /// actual arguments.  Lower distance = more specific match.
-    fn method_candidate_type_distance(&self, args: &[Value], def: &MethodDef) -> usize {
-        let mut total = 0usize;
-        let mut arg_idx = 0;
-        for pd in &def.param_defs {
-            if pd.is_invocant || pd.named || pd.name.starts_with("__type_capture__") {
-                continue;
-            }
-            // Slurpy parameters (`*@x`, `*%h`, `**@x`) are less specific than a
-            // required positional of the same type. When a single Positional
-            // argument matches both `(@x)` and `(*@x)`, the non-slurpy candidate
-            // must win rather than being treated as equally specific (ambiguous).
-            if pd.slurpy || pd.double_slurpy {
-                total += 2000;
-                arg_idx += 1;
-                continue;
-            }
-            if let Some(tc) = &pd.type_constraint {
-                let base = Self::constraint_base_for_distance(tc);
-                if arg_idx < args.len() {
-                    total += Self::builtin_type_distance(base, &args[arg_idx]);
-                }
-            } else {
-                total += 1000;
-            }
-            arg_idx += 1;
-        }
-        total
-    }
-
-    fn constraint_base_for_distance(constraint: &str) -> &str {
-        let s = if constraint.ends_with(":D") || constraint.ends_with(":U") {
-            &constraint[..constraint.len() - 2]
-        } else {
-            constraint
-        };
-        s.split('(').next().unwrap_or(s)
-    }
-
-    /// Compute the type hierarchy distance between a constraint and a value.
-    /// 0 = exact match, larger = less specific.
-    fn builtin_type_distance(constraint: &str, value: &Value) -> usize {
-        let value_type = super::value_type_name(value);
-        if constraint == value_type {
-            return 0;
-        }
-        if let Value::Instance { class_name, .. } = value
-            && constraint == class_name.resolve().as_str()
-        {
-            return 0;
-        }
-        let builtin_mro: &[&str] = match value_type {
-            "Bool" => &["Bool", "Int", "Numeric", "Real", "Cool", "Any", "Mu"],
-            "Int" => &["Int", "Numeric", "Real", "Cool", "Any", "Mu"],
-            "Num" => &["Num", "Numeric", "Real", "Cool", "Any", "Mu"],
-            "Rat" | "FatRat" => &["Rat", "Numeric", "Real", "Cool", "Any", "Mu"],
-            "Complex" => &["Complex", "Numeric", "Cool", "Any", "Mu"],
-            "Str" => &["Str", "Stringy", "Cool", "Any", "Mu"],
-            "Array" => &["Array", "List", "Positional", "Cool", "Any", "Mu"],
-            "List" => &["List", "Positional", "Cool", "Any", "Mu"],
-            "Hash" => &["Hash", "Map", "Associative", "Cool", "Any", "Mu"],
-            "Pair" => &["Pair", "Associative", "Cool", "Any", "Mu"],
-            "Range" => &["Range", "Positional", "Cool", "Any", "Mu"],
-            "Set" => &["Set", "Setty", "QuantHash", "Associative", "Any", "Mu"],
-            "Bag" => &["Bag", "Baggy", "QuantHash", "Associative", "Any", "Mu"],
-            "Mix" => &[
-                "Mix",
-                "Mixy",
-                "Baggy",
-                "QuantHash",
-                "Associative",
-                "Any",
-                "Mu",
-            ],
-            "Sub" => &["Sub", "Routine", "Block", "Code", "Callable", "Any", "Mu"],
-            "Seq" => &["Seq", "Positional", "Cool", "Any", "Mu"],
-            "Regex" => &[
-                "Regex", "Method", "Routine", "Block", "Code", "Callable", "Any", "Mu",
-            ],
-            "Junction" => &["Junction", "Mu"],
-            _ => &[],
-        };
-        for (i, &ancestor) in builtin_mro.iter().enumerate() {
-            if ancestor == constraint {
-                return i;
-            }
-        }
-        500
-    }
-
-    pub(crate) fn resolve_all_methods_with_owner(
-        &mut self,
-        class_name: &str,
-        method_name: &str,
-        arg_values: &[Value],
-    ) -> Vec<(String, MethodDef)> {
-        let role_bindings = self.registry().get_role_param_bindings(class_name);
-        let mro = self.class_mro(class_name);
-        let mut matches = Vec::new();
-        for cn in &mro {
-            // An MRO entry may be a regular class or a punned role used as a
-            // parent (`class Foo is R1 { ... }`). Role methods are stored in
-            // `self.registry().roles`, so fall back there when the entry is not a class.
-            // Single guard for both the class and role method tables; clone the
-            // matched overload Vec out so the guard drops before re-entry below.
-            let overloads = {
-                let registry = self.registry();
-                registry
-                    .classes
-                    .get(cn.as_str())
-                    .and_then(|c| c.methods.get(method_name))
-                    .or_else(|| {
-                        registry
-                            .roles
-                            .get(cn.as_str())
-                            .and_then(|r| r.methods.get(method_name))
-                    })
-                    .cloned()
-            };
-            if let Some(overloads) = overloads {
-                let is_ancestor = cn != class_name;
-                for def in overloads {
-                    if def.is_private {
-                        continue;
-                    }
-                    // Submethods are NOT inherited
-                    if def.is_my && is_ancestor {
-                        continue;
-                    }
-                    if self.method_args_match_for_invocant(
-                        class_name,
-                        &def,
-                        arg_values,
-                        role_bindings.as_ref(),
-                        None,
-                    ) {
-                        matches.push((cn.clone(), def));
-                    }
-                }
-            }
-        }
-        matches
-    }
-
-    /// Resolve methods for .*/,+ dispatch: one resolution per MRO level.
-    /// For multi methods, dispatches through the proto at each level.
-    /// Returns empty vec if any level has the method but no candidate matches.
-    pub(crate) fn resolve_methods_per_mro_level(
-        &mut self,
-        class_name: &str,
-        method_name: &str,
-        arg_values: &[Value],
-    ) -> Vec<(String, MethodDef)> {
-        let mro = self.class_mro(class_name);
-        let mut defining_levels: Vec<String> = Vec::new();
-        for cn in &mro {
-            let is_ancestor = cn != class_name;
-            if let Some(overloads) = self
-                .registry()
-                .get_method_overloads(cn.as_str(), method_name)
-            {
-                let has_visible = overloads
-                    .iter()
-                    .any(|d| !d.is_private && (!d.is_my || !is_ancestor));
-                if has_visible {
-                    defining_levels.push(cn.clone());
-                }
-            }
-        }
-        if defining_levels.is_empty() {
-            return Vec::new();
-        }
-        let any_multi = defining_levels.iter().any(|cn| {
-            self.registry()
-                .get_method_overloads(cn.as_str(), method_name)
-                .is_some_and(|ovs| ovs.iter().any(|d| d.is_multi))
-        });
-        if !any_multi {
-            // Non-multi: use the standard resolution
-            return self.resolve_all_methods_with_owner(class_name, method_name, arg_values);
-        }
-        // Multi methods: dispatch through proto at each level
-        let mut matches = Vec::new();
-        let mut any_failed = false;
-        for cn in &defining_levels {
-            if let Some(resolved) =
-                self.resolve_method_with_owner_impl(cn, method_name, arg_values, None)
-            {
-                matches.push(resolved);
-            } else {
-                any_failed = true;
-            }
-        }
-        if any_failed {
-            return Vec::new();
-        }
-        matches
-    }
-
-    pub(crate) fn should_skip_defer_method_candidate(
-        &self,
-        receiver_class: &str,
-        candidate_owner: &str,
-    ) -> bool {
-        if receiver_class != candidate_owner && self.registry().is_hidden_class(candidate_owner) {
-            return true;
-        }
-        self.registry()
-            .is_hidden_defer_parent(receiver_class, candidate_owner)
-    }
-
-    pub(super) fn resolve_private_method_with_owner(
-        &mut self,
-        class_name: &str,
-        owner_class: &str,
-        method_name: &str,
-        arg_values: &[Value],
-    ) -> Option<(String, MethodDef)> {
-        let role_bindings = self.registry().get_role_param_bindings(class_name);
-        let mro = self.class_mro(class_name);
-        for cn in mro {
-            if cn != owner_class {
-                continue;
-            }
-            // Hoist clone to a `let` so the guard drops before re-entry (&mut self).
-            let overloads = self.registry().get_method_overloads(&cn, method_name);
-            if let Some(overloads) = overloads {
-                for def in overloads {
-                    if !def.is_private {
-                        continue;
-                    }
-                    if self.method_args_match_for_invocant(
-                        class_name,
-                        &def,
-                        arg_values,
-                        role_bindings.as_ref(),
-                        None,
-                    ) {
-                        return Some((cn.clone(), def));
-                    }
-                }
-            }
-        }
-        None
-    }
-
-    pub(super) fn resolve_private_method_any_owner(
-        &mut self,
-        class_name: &str,
-        method_name: &str,
-        arg_values: &[Value],
-    ) -> Option<(String, MethodDef)> {
-        let role_bindings = self.registry().get_role_param_bindings(class_name);
-        if arg_values.is_empty()
-            && let Some(cached) = self
-                .private_zeroarg_method_cache
-                .get(&(class_name.to_string(), method_name.to_string()))
-        {
-            return cached.clone();
-        }
-        let mro = self.class_mro(class_name);
-        // Fast path: when there are no positional args, avoid cloning the
-        // overloads vector by scanning with a shared borrow first. This covers
-        // the common case of zero-argument private method calls in tight loops.
-        if arg_values.is_empty() {
-            // Scan with a shared registry borrow (avoids cloning the whole
-            // overloads Vec — only the single matched def is cloned), find the
-            // candidate, then drop the guard before mutating the cache.
-            let mut resolved: Option<(String, MethodDef)> = None;
-            'scan: for cn in &mro {
-                let registry = self.registry();
-                if let Some(overloads) = registry
-                    .classes
-                    .get(cn)
-                    .and_then(|c| c.methods.get(method_name))
-                {
-                    // First pass: skip stubs
-                    for def in overloads {
-                        if !def.is_private {
-                            continue;
-                        }
-                        if Self::is_stub_method_body(&def.body) {
-                            continue;
-                        }
-                        if def
-                            .param_defs
-                            .iter()
-                            .all(|p| p.is_invocant || p.traits.iter().any(|t| t == "invocant"))
-                        {
-                            resolved = Some((cn.clone(), def.clone()));
-                            break 'scan;
-                        }
-                    }
-                    // Second pass: include stubs
-                    for def in overloads {
-                        if !def.is_private {
-                            continue;
-                        }
-                        if def
-                            .param_defs
-                            .iter()
-                            .all(|p| p.is_invocant || p.traits.iter().any(|t| t == "invocant"))
-                        {
-                            resolved = Some((cn.clone(), def.clone()));
-                            break 'scan;
-                        }
-                    }
-                }
-            }
-            if let Some(resolved) = resolved {
-                self.private_zeroarg_method_cache.insert(
-                    (class_name.to_string(), method_name.to_string()),
-                    Some(resolved.clone()),
-                );
-                return Some(resolved);
-            }
-        }
-        for cn in mro {
-            // Hoist clone to a `let` so the guard drops before re-entry (&mut self).
-            let overloads = self.registry().get_method_overloads(&cn, method_name);
-            if let Some(overloads) = overloads {
-                for def in &overloads {
-                    if !def.is_private {
-                        continue;
-                    }
-                    if Self::is_stub_method_body(&def.body) {
-                        continue;
-                    }
-                    if self.method_args_match_for_invocant(
-                        class_name,
-                        def,
-                        arg_values,
-                        role_bindings.as_ref(),
-                        None,
-                    ) {
-                        return Some((cn.clone(), def.clone()));
-                    }
-                }
-                for def in overloads {
-                    if !def.is_private {
-                        continue;
-                    }
-                    if self.method_args_match_for_invocant(
-                        class_name,
-                        &def,
-                        arg_values,
-                        role_bindings.as_ref(),
-                        None,
-                    ) {
-                        return Some((cn.clone(), def));
-                    }
-                }
-            }
-        }
-        None
-    }
-
-    pub(crate) fn resolve_private_method_for_vm(
-        &mut self,
-        class_name: &str,
-        method: &str,
-        arg_values: &[Value],
-    ) -> Option<(String, MethodDef)> {
-        let private_rest = method.strip_prefix('!')?;
-        if let Some((owner_class, pm_name)) = private_rest.split_once("::") {
-            self.resolve_private_method_with_owner(class_name, owner_class, pm_name, arg_values)
-        } else {
-            self.resolve_private_method_any_owner(class_name, private_rest, arg_values)
-        }
-    }
-
-    pub(crate) fn can_fast_dispatch_private_method_vm(&self, owner_class: &str) -> bool {
-        self.method_class_stack
-            .last()
-            .is_some_and(|caller| caller == owner_class)
-    }
-
-    /// Resolve the MRO for `class_name`. Delegates to [`Registry::class_mro`],
-    /// which holds a single write guard for the whole compute-and-cache op.
-    pub(crate) fn class_mro(&mut self, class_name: &str) -> Vec<String> {
-        self.registry_mut().class_mro(class_name)
-    }
-
-    pub(super) fn force_lazy_list(&mut self, list: &LazyList) -> Result<Vec<Value>, RuntimeError> {
-        // A lazy map/grep pipeline is rooted at an infinite source. It can still
-        // terminate when the callback runs `last`; attempt a bounded force and
-        // return the result if the pipe became `done`, otherwise throw
-        // X::Cannot::Lazy (genuinely infinite). Mirrors `force_lazy_list_vm`.
-        if list.lazy_pipe.is_some() {
-            const EAGER_FORCE_CAP: usize = 1_000_000;
-            let forced = self.force_lazy_pipe(list, EAGER_FORCE_CAP)?;
-            let done = list
-                .lazy_pipe
-                .as_ref()
-                .map(|p| p.lock().unwrap().done)
-                .unwrap_or(true);
-            if done {
-                return Ok(forced);
-            }
-            return Err(RuntimeError::typed_msg(
-                "X::Cannot::Lazy",
-                "Cannot coerce an infinite lazy list to a strict list",
-            ));
-        }
-        if let Some(cached) = list.cache.lock().unwrap().clone() {
-            return Ok(cached);
-        }
-        // Lazy map: evaluate the stored callback over the stored items now.
-        //
-        // `.map` returns a lazy Seq in Raku, so the callback runs only when the
-        // Seq is forced. By deferring evaluation to this point we get the
-        // correct out-of-dynamic-scope behaviour for `return` inside the
-        // callback for free: if the lexically enclosing routine has already
-        // exited (it is no longer on the dynamic call stack), the existing
-        // CX::Return propagation machinery (calls.rs / the top-level VM loop)
-        // converts the signal into `X::ControlFlow::Return` with
-        // out-of-dynamic-scope set. Delegating to `eval_map_over_items` keeps
-        // full fidelity with eager map: block arity > 1, Slip flattening,
-        // LAST/NEXT phasers, and composed callbacks all behave identically.
-        if let (Some(map_items), Some(map_func)) = (
-            list.env.get("__mutsu_lazy_map_items"),
-            list.env.get("__mutsu_lazy_map_func"),
-        ) {
-            let items_list = crate::runtime::value_to_list(map_items);
-            let func = map_func.clone();
-            let result = self.eval_map_over_items(Some(func), items_list)?;
-            let result_items = crate::runtime::value_to_list(&result);
-            *list.cache.lock().unwrap() = Some(result_items.clone());
-            return Ok(result_items);
-        }
-        let saved_env = self.env.clone();
-        let saved_len = self.gather_items.len();
-        self.env = list.env.clone();
-        self.gather_items.push(Vec::new());
-        self.gather_take_limits.push(None);
-        let run_res = self.run_block(&list.body);
-        let items = self.gather_items.pop().unwrap_or_default();
-        self.gather_take_limits.pop();
-        while self.gather_items.len() > saved_len {
-            self.gather_items.pop();
-            self.gather_take_limits.pop();
-        }
-        let mut merged_env = saved_env;
-        for (k, v) in self.env.iter() {
-            merged_env.insert_sym(*k, v.clone());
-        }
-        self.env = merged_env;
-        run_res?;
-        *list.cache.lock().unwrap() = Some(items.clone());
-        Ok(items)
-    }
-
-    pub(crate) fn force_lazy_list_prefix_bridge(
-        &mut self,
-        list: &crate::value::LazyList,
-        needed_len: usize,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        if needed_len == 0 {
-            return Ok(Vec::new());
-        }
-        if let Some(cached) = list.cache.lock().unwrap().clone()
-            && cached.len() >= needed_len
-        {
-            return Ok(cached[..needed_len].to_vec());
-        }
-
-        let saved_env = self.env.clone();
-        let saved_len = self.gather_items.len();
-        self.env = list.env.clone();
-        self.gather_items.push(Vec::new());
-        self.gather_take_limits.push(Some(needed_len));
-        let run_res = self.run_block(&list.body);
-        let mut items = self.gather_items.pop().unwrap_or_default();
-        self.gather_take_limits.pop();
-        while self.gather_items.len() > saved_len {
-            self.gather_items.pop();
-            self.gather_take_limits.pop();
-        }
-        let mut merged_env = saved_env;
-        for (k, v) in self.env.iter() {
-            merged_env.insert_sym(*k, v.clone());
-        }
-        self.env = merged_env;
-
-        match run_res {
-            Ok(()) => {
-                *list.cache.lock().unwrap() = Some(items.clone());
-            }
-            Err(err) if err.message == Self::LAZY_GATHER_TAKE_LIMIT_SIGNAL => {}
-            Err(err) => return Err(err),
-        }
-        if items.len() > needed_len {
-            items.truncate(needed_len);
-        }
-        Ok(items)
-    }
-
-    pub(crate) fn force_lazy_list_bridge(
-        &mut self,
-        list: &crate::value::LazyList,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        self.force_lazy_list(list)
-    }
-
-    /// Force a `LazyIoLines` value into an eager array by reading all remaining
-    /// records from the file handle. When `words` is true the records are
-    /// whitespace-delimited words; otherwise they are lines.
-    pub(crate) fn force_lazy_io_lines(
-        &mut self,
-        handle: &Value,
-        words: bool,
-    ) -> Result<Value, RuntimeError> {
-        let mut items = Vec::new();
-        if words {
-            while let Some(word) = self.read_word_from_handle_value(handle)? {
-                items.push(Value::str(word));
-            }
-        } else {
-            while let Some(line) = self.read_line_from_handle_value(handle)? {
-                items.push(Value::str(line));
-            }
-        }
-        Ok(Value::array(items))
-    }
-
-    /// Force a `LazyIoLines` value into an eager array by reading at most `n`
-    /// records from the file handle, leaving the rest available for subsequent
-    /// reads. When `words` is true the records are words; otherwise lines.
-    pub(crate) fn force_lazy_io_lines_n(
-        &mut self,
-        handle: &Value,
-        n: usize,
-        words: bool,
-    ) -> Result<Value, RuntimeError> {
-        let mut items = Vec::new();
-        while items.len() < n {
-            let next = if words {
-                self.read_word_from_handle_value(handle)?
-            } else {
-                self.read_line_from_handle_value(handle)?
-            };
-            match next {
-                Some(rec) => items.push(Value::str(rec)),
-                None => break,
-            }
-        }
-        Ok(Value::array(items))
-    }
-
-    pub(super) fn split_block_phasers(&self, stmts: &[Stmt]) -> SplitPhasers {
-        let mut pre_ph = Vec::new();
-        let mut enter_ph = Vec::new();
-        let mut body_main = Vec::new();
-        let mut success_queue = Vec::new();
-        let mut failure_queue = Vec::new();
-        let mut post_ph = Vec::new();
-        for stmt in stmts {
-            if let Stmt::Phaser { kind, body } = stmt {
-                match kind {
-                    PhaserKind::Pre => pre_ph.push(Stmt::Block(body.clone())),
-                    PhaserKind::Enter => enter_ph.push(Stmt::Block(body.clone())),
-                    PhaserKind::Post | PhaserKind::Leave | PhaserKind::Keep | PhaserKind::Undo => {}
-                    _ => body_main.push(stmt.clone()),
-                }
-            } else {
-                body_main.push(stmt.clone());
-            }
-        }
-        // POST phasers in reverse source order
-        for stmt in stmts.iter().rev() {
-            if let Stmt::Phaser { kind, body } = stmt {
-                match kind {
-                    PhaserKind::Leave => {
-                        success_queue.push(Stmt::Block(body.clone()));
-                        failure_queue.push(Stmt::Block(body.clone()));
-                    }
-                    PhaserKind::Keep => success_queue.push(Stmt::Block(body.clone())),
-                    PhaserKind::Undo => failure_queue.push(Stmt::Block(body.clone())),
-                    PhaserKind::Post => post_ph.push(Stmt::Block(body.clone())),
-                    _ => {}
-                }
-            }
-        }
-        (
-            pre_ph,
-            enter_ph,
-            success_queue,
-            failure_queue,
-            post_ph,
-            body_main,
-        )
-    }
-
-    pub(super) fn make_promise_instance(&self, status: &str, result: Value) -> Value {
-        let mut attrs = HashMap::new();
-        attrs.insert("status".to_string(), Value::str(status.to_string()));
-        attrs.insert("result".to_string(), result);
-        Value::make_instance(Symbol::intern("Promise"), attrs)
-    }
-
-    pub(super) fn make_supply_instance(&self) -> Value {
-        let sid = super::native_methods::next_supply_id();
-        let mut attrs = HashMap::new();
-        attrs.insert("values".to_string(), Value::array(Vec::new()));
-        attrs.insert("taps".to_string(), Value::array(Vec::new()));
-        attrs.insert("supply_id".to_string(), Value::Int(sid as i64));
-        Value::make_instance(Symbol::intern("Supply"), attrs)
-    }
-
-    pub(crate) fn call_sub_value(
-        &mut self,
-        func: Value,
-        args: Vec<Value>,
-        merge_all: bool,
-    ) -> Result<Value, RuntimeError> {
-        // Upgrade WeakSub to Sub transparently
-        let func = match func {
-            Value::WeakSub(ref weak) => match weak.upgrade() {
-                Some(strong) => Value::Sub(strong),
-                None => return Err(RuntimeError::new("Callable has been freed")),
-            },
-            other => other,
-        };
-        if let Value::Routine { package, name, .. } = &func {
-            if !package.is_empty() && package != "GLOBAL" {
-                let fq = format!("{package}::{name}");
-                if self.resolve_function(&fq).is_some() {
-                    return self.call_function(&fq, args);
-                }
-            }
-            let name_str = name.resolve();
-            if self.resolve_function(&name_str).is_some()
-                || self.has_proto(&name_str)
-                || self.has_multi_candidates(&name_str)
-            {
-                return self.call_function(&name_str, args);
-            }
-            // Method dispatch fallback for &?ROUTINE.dispatcher()(self, ...)
-            // Only use this when the package is a known class.
-            let pkg = package.resolve();
-            if !args.is_empty() && !pkg.is_empty() && pkg != "GLOBAL" && self.has_class(&pkg) {
-                let invocant = args[0].clone();
-                let method_args = args[1..].to_vec();
-                return self.call_method_with_values(invocant, &name_str, method_args);
-            }
-            return self.call_function(&name_str, args);
-        }
-        if let Value::Junction { kind, values } = func {
-            let mut results = Vec::with_capacity(values.len());
-            for callable in values.iter() {
-                results.push(self.call_sub_value(callable.clone(), args.clone(), merge_all)?);
-            }
-            return Ok(Value::junction(kind, results));
-        }
-        if let Value::Sub(data) = func {
-            // Check for wrap chain — if wrappers exist, dispatch through them
-            // Skip if we're already inside a wrap dispatch for this sub
-            let already_dispatching = self.wrap_dispatch_stack.iter().any(|f| f.sub_id == data.id);
-            if !already_dispatching
-                && let Some(chain) = self.wrap_chains.get(&data.id).cloned()
-                && !chain.is_empty()
-            {
-                let (sanitized_args, callsite_line) = self.sanitize_call_args(&args);
-                self.test_pending_callsite_line = callsite_line;
-                // Build remaining list: inner wrappers then original sub
-                // chain is ordered inner-to-outer (last = outermost), so:
-                // outermost is last, we call it; remaining = [inner..., original]
-                let outermost = chain.last().unwrap().1.clone();
-                let mut remaining: Vec<Value> = Vec::new();
-                // Add wrappers from second-to-last down to first (inner order)
-                for i in (0..chain.len() - 1).rev() {
-                    remaining.push(chain[i].1.clone());
-                }
-                // Add the original sub at the end
-                remaining.push(Value::Sub(data.clone()));
-                let frame = super::WrapDispatchFrame {
-                    sub_id: data.id,
-                    remaining,
-                    args: sanitized_args.clone(),
-                };
-                let (wrapper_id, wrapper_base_env) = if let Value::Sub(ref wd) = outermost {
-                    (Some(wd.id), Some(wd.env.clone()))
-                } else {
-                    (None, None)
-                };
-                self.wrap_dispatch_stack.push(frame);
-                let result = self.call_sub_value(outermost, sanitized_args, false);
-                self.wrap_dispatch_stack.pop();
-                // Propagate closure variable mutations from the wrapper back to
-                // the current env so captured variables (e.g. $seen = True) are
-                // visible to the caller. Only write back keys the wrapper actually
-                // *changed* relative to its captured lexical snapshot: a persisted
-                // value equal to the wrapper's capture is stale (never mutated) and
-                // must not clobber the caller's live value (e.g. the `$h` that holds
-                // this very WrapHandle, captured as Nil mid-`my $h = &foo.wrap(...)`).
-                if let Some(wid) = wrapper_id
-                    && let Some(persisted) = self.closure_env_overrides.get(&wid).cloned()
-                {
-                    for (k, v) in persisted.iter() {
-                        if !self.env.contains_key_sym(*k) {
-                            continue;
-                        }
-                        let unchanged = wrapper_base_env
-                            .as_ref()
-                            .is_some_and(|base| base.get_sym(*k) == Some(v));
-                        if unchanged {
-                            continue;
-                        }
-                        self.env.insert_sym(*k, v.clone());
-                    }
-                }
-                return result;
-            }
-            let (sanitized_args, callsite_line) = self.sanitize_call_args(&args);
-            self.test_pending_callsite_line = callsite_line;
-            let mut call_args = sanitized_args.clone();
-            if !data.assumed_positional.is_empty() || !data.assumed_named.is_empty() {
-                let mut positional = Vec::new();
-                let mut named = data.assumed_named.clone();
-                let mut incoming_positional = Vec::new();
-                for arg in &sanitized_args {
-                    if let Value::Pair(key, boxed) = arg {
-                        named.insert(key.clone(), *boxed.clone());
-                    } else {
-                        incoming_positional.push(arg.clone());
-                    }
-                }
-                // Fill bare `*` primers from incoming positional args in order.
-                // Remaining positional args are appended after all fixed primers.
-                let mut incoming_idx = 0usize;
-                for assumed in &data.assumed_positional {
-                    let is_placeholder = matches!(assumed, Value::Whatever)
-                        || matches!(assumed, Value::Num(f) if f.is_infinite())
-                        || matches!(assumed, Value::Rat(_, 0));
-                    if is_placeholder {
-                        if incoming_idx < incoming_positional.len() {
-                            positional.push(incoming_positional[incoming_idx].clone());
-                            incoming_idx += 1;
-                        }
-                    } else {
-                        positional.push(assumed.clone());
-                    }
-                }
-                positional.extend(incoming_positional.into_iter().skip(incoming_idx));
-                call_args = positional;
-                for (key, value) in named {
-                    call_args.push(Value::Pair(key, Box::new(value)));
-                }
-            }
-            // Routine wrapper from .assuming() on a multi-dispatch sub
-            if let Some(Value::Str(routine_name)) = data.env.get("__mutsu_routine_name").cloned() {
-                return self.call_function(&routine_name, call_args);
-            }
-            // Multi-dispatch dispatcher: captured multi candidates from resolve_code_var
-            if let Some(Value::Array(candidates_arc, _)) =
-                data.env.get("__mutsu_multi_dispatch_candidates").cloned()
-            {
-                let candidates = (*candidates_arc).clone();
-                // First try to dispatch via the function table (if still in scope)
-                if let Some(Value::Str(name)) = data.env.get("__mutsu_multi_dispatch_name").cloned()
-                    && (self.resolve_function(&name).is_some()
-                        || self.has_proto(&name)
-                        || self.has_multi_candidates(&name))
-                {
-                    return self.call_function(&name, call_args);
-                }
-                // Candidates are out of scope -- dispatch through captured Subs
-                for candidate in &candidates {
-                    if let Value::Sub(cand_data) = candidate
-                        && self
-                            .bind_function_args_values(
-                                &cand_data.param_defs,
-                                &cand_data.params,
-                                &call_args,
-                            )
-                            .is_ok()
-                    {
-                        return self.call_sub_value(candidate.clone(), call_args, false);
-                    }
-                }
-                // Slurpy catch-all
-                for candidate in &candidates {
-                    if let Value::Sub(cand_data) = candidate
-                        && cand_data.param_defs.iter().any(|pd| pd.slurpy)
-                    {
-                        return self.call_sub_value(candidate.clone(), call_args, false);
-                    }
-                }
-                if let Some(candidate) = candidates.first() {
-                    return self.call_sub_value(candidate.clone(), call_args, false);
-                }
-            }
-            if let (Some(left), Some(right)) = (
-                data.env.get("__mutsu_compose_left").cloned(),
-                data.env.get("__mutsu_compose_right").cloned(),
-            ) {
-                let right_result = self.call_sub_value(right, call_args, false)?;
-                let (left_params, left_param_defs) = self.callable_signature(&left);
-                let left_variadic = left_param_defs.iter().any(|pd| pd.slurpy && !pd.named);
-                let left_expects_single = match &left {
-                    Value::Sub(left_data) if left_params.is_empty() => {
-                        let (uses_positional, _) = Self::auto_signature_uses(&left_data.body);
-                        !uses_positional
-                    }
-                    _ => !left_variadic && left_params.len() == 1,
-                };
-                let left_args = Self::composed_result_to_args(right_result, left_expects_single);
-                return self.call_sub_value(left, left_args, false);
-            }
-            // Reject arguments for blocks with explicitly empty signatures
-            // (e.g. `-> { ... }` which takes 0 parameters).
-            if data.empty_sig && !call_args.is_empty() {
-                return Err(Self::reject_args_for_empty_sig(&call_args));
-            }
-            let saved_env = self.env.clone();
-            let saved_readonly = self.save_readonly_vars();
-            if let Some(line) = self.test_pending_callsite_line {
-                self.env.insert("?LINE".to_string(), Value::Int(line));
-            }
-            self.push_caller_env();
-            let persist_closure_env =
-                data.name.is_empty() || !self.has_function(&data.name.resolve());
-            let closure_base_env = if persist_closure_env {
-                self.closure_env_overrides
-                    .get(&data.id)
-                    .cloned()
-                    .unwrap_or_else(|| data.env.clone())
-            } else {
-                data.env.clone()
-            };
-            let mut new_env = saved_env.clone();
-            for (k, v) in &closure_base_env {
-                // A captured `ContainerRef` is a shared container cell (box-on-
-                // capture / lever C): the single source of truth for that lexical,
-                // so it must OVERWRITE any stale value the caller env holds (a
-                // value the declaring block leaked, or this closure's own prior
-                // writeback). The default "don't overwrite" merge would hide the
-                // live cell and make a *second* bareword `f()` call read a stale
-                // value (`my &f; { my $a=3; &f=sub{$a++} }; f(); f()` returned 3
-                // then 0). Mirrors the VM closure dispatch (`call_compiled_closure`).
-                if matches!(v, Value::ContainerRef(_)) {
-                    new_env.insert_sym(*k, v.clone());
-                    continue;
-                }
-                if merge_all {
-                    new_env.entry_or_insert(k.resolve(), v.clone());
-                    continue;
-                }
-                if matches!(new_env.get_sym(*k), Some(Value::Array(..)))
-                    && matches!(v, Value::Array(..))
-                {
-                    continue;
-                }
-                new_env.insert_sym(*k, v.clone());
-            }
-            self.env = new_env.clone();
-            // Check for empty signature: a pointy block `-> { }` or sub with no
-            // params that doesn't use @_/%_ must reject positional arguments.
-            let positional_call_args: Vec<&Value> = call_args
-                .iter()
-                .filter(|v| !matches!(v, Value::Pair(_, _)))
-                .collect();
-            if data.empty_sig && !positional_call_args.is_empty() {
-                self.pop_caller_env();
-                self.env = saved_env;
-                self.restore_readonly_vars(saved_readonly);
-                return Err(Self::reject_args_for_empty_sig(&call_args));
-            }
-            let rw_bindings =
-                match self.bind_function_args_values(&data.param_defs, &data.params, &call_args) {
-                    Ok(bindings) => bindings,
-                    Err(e) => {
-                        self.pop_caller_env();
-                        self.env = saved_env;
-                        self.restore_readonly_vars(saved_readonly);
-                        return Err(e);
-                    }
-                };
-            new_env = self.env.clone();
-            if data.params.is_empty() {
-                for arg in &sanitized_args {
-                    if let Value::Pair(name, value) = arg {
-                        new_env.insert(format!(":{}", name), *value.clone());
-                    }
-                }
-            }
-            // Bind implicit $_ for bare blocks called with arguments
-            let (uses_positional, _) = Self::auto_signature_uses(&data.body);
-            if data.params.is_empty()
-                && !uses_positional
-                && !sanitized_args.is_empty()
-                && let Some(first_positional) = sanitized_args
-                    .iter()
-                    .find(|v| !matches!(v, Value::Pair(_, _)))
-            {
-                new_env.insert("_".to_string(), first_positional.clone());
-                new_env.insert("$_".to_string(), first_positional.clone());
-            } else if data.params.is_empty()
-                && sanitized_args.is_empty()
-                && data.name.is_empty()
-                && let Some(caller_topic) = saved_env.get("_")
-            {
-                new_env.insert("_".to_string(), caller_topic.clone());
-                new_env.insert("$_".to_string(), caller_topic.clone());
-            }
-            // &?BLOCK: weak self-reference to break reference cycles
-            let block_arc = std::sync::Arc::new(crate::value::SubData {
-                package: data.package,
-                name: data.name,
-                params: data.params.clone(),
-                param_defs: data.param_defs.clone(),
-                body: data.body.clone(),
-                is_rw: data.is_rw,
-                is_raw: data.is_raw,
-                env: new_env.clone(),
-                assumed_positional: data.assumed_positional.clone(),
-                assumed_named: data.assumed_named.clone(),
-                id: data.id,
-                empty_sig: data.empty_sig,
-                is_bare_block: data.is_bare_block,
-                compiled_code: data.compiled_code.clone(),
-                deprecated_message: data.deprecated_message.clone(),
-                source_line: data.source_line,
-                source_file: data.source_file.clone(),
-                owned_captures: data.owned_captures.clone(),
-                upvalues: data.upvalues.clone(),
-            });
-            new_env.insert(
-                "&?BLOCK".to_string(),
-                Value::WeakSub(std::sync::Arc::downgrade(&block_arc)),
-            );
-            let block_sub = Value::make_sub_with_id(
-                data.package,
-                data.name,
-                vec![],
-                Vec::new(),
-                data.body.clone(),
-                data.is_rw,
-                new_env.clone(),
-                data.id,
-            );
-            self.env = new_env;
-            self.env.insert(
-                "__mutsu_callable_id".to_string(),
-                Value::Int(data.id as i64),
-            );
-            self.routine_stack.push(RoutineFrame {
-                package: data.package.resolve(),
-                name: data.name.resolve(),
-                line: None,
-                file: None,
-                is_method: false,
-                is_block: true,
-            });
-            self.block_stack.push(block_sub);
-            let return_spec = data.env.get("__mutsu_return_type").and_then(|v| match v {
-                Value::Str(s) => Some(s.to_string()),
-                _ => None,
-            });
-            self.prepare_definite_return_slot(return_spec.as_deref());
-            let let_mark = self.let_saves.len();
-            // Snapshot the body-entry env so the exit writeback can tell which
-            // captured outer scalars the body actually *mutated* (and propagate
-            // only those) from those it merely captured a stale snapshot of.
-            // Without this, a callee's stale captured value (e.g. an original
-            // sub's snapshot of `$h` taken before `my $h = &foo.wrap(...)`
-            // assigned it) would clobber the caller's live value on return.
-            // Env is copy-on-write, so this clone is O(1) until the body forks it.
-            let body_entry_env = self.env.clone();
-            let result = match self.eval_block_value(&data.body) {
-                Err(mut e) if e.is_leave => {
-                    let routine_key = format!("{}::{}", data.package, data.name);
-                    let matches_frame = if let Some(target_id) = e.leave_callable_id {
-                        target_id == data.id
-                    } else if let Some(target_routine) = e.leave_routine.as_ref() {
-                        target_routine == &routine_key
-                    } else {
-                        e.label.is_none()
-                    };
-                    if matches_frame {
-                        e.is_leave = false;
-                        e.control = None;
-                        if e.return_value.is_none() {
-                            e.return_value = Some(Value::Nil);
-                        }
-                        Err(e)
-                    } else {
-                        Err(e)
-                    }
-                }
-                other => other,
-            };
-            let effective_return_spec = return_spec
-                .as_deref()
-                .map(|spec| self.resolved_type_capture_name(spec));
-            self.block_stack.pop();
-            self.routine_stack.pop();
-            // Manage let saves based on sub result
-            match &result {
-                Ok(_) => {
-                    // Successful completion — restore temps, discard lets
-                    self.resolve_let_saves_on_success(let_mark, true);
-                }
-                Err(e) if e.return_value.is_some() => {
-                    // Explicit return — restore temps, discard lets
-                    self.resolve_let_saves_on_success(let_mark, true);
-                }
-                Err(_) => {
-                    // Exception/fail — restore saves
-                    self.restore_let_saves(let_mark);
-                }
-            }
-            if persist_closure_env {
-                let mut persisted_closure_env = closure_base_env.clone();
-                for key in closure_base_env.keys() {
-                    if let Some(value) = self.env.get_sym(*key).cloned() {
-                        persisted_closure_env.insert_sym(*key, value);
-                    }
-                }
-                self.closure_env_overrides
-                    .insert(data.id, persisted_closure_env);
-            }
-            // Preserve map rw topic writeback across env restoration
-            let rw_map_topic = self.env.get("__mutsu_rw_map_topic__").cloned();
-            let mut merged = saved_env;
-            self.pop_caller_env_with_writeback(&mut merged);
-            // Slice 1b (captured-outer cell sharing): a nested-declared method/sub
-            // dispatched through `call_sub_value` (`$d.bar` where `method bar { … }`
-            // was declared in `foo`'s body) runs its body via the interpreter and
-            // merges a captured-outer scalar write back into `merged` (the caller
-            // env) below. But the caller's matching *local slot* is not refreshed
-            // from env unless the blanket reconcile is on. Record each captured-outer
-            // scalar this body actually *changed* (vs its body-entry snapshot) so the
-            // call site (`apply_pending_rw_writeback`) writes the merged value straight
-            // through to the caller's slot, dropping the dependency on the reverse
-            // `sync_locals_from_env` pull. Mirrors `merge_method_env`'s
-            // `changed_caller_locals` for the compiled method path.
-            let mut captured_outer_writes: Vec<String> = Vec::new();
-            if merge_all {
-                for (k, v) in self.env.iter() {
-                    if k != "_"
-                        && k != "@_"
-                        && (merged.contains_key_sym(*k) || k.starts_with("__mutsu_var_meta::"))
-                    {
-                        if merged.contains_key_sym(*k)
-                            && matches!(
-                                v,
-                                Value::Bool(_)
-                                    | Value::Int(_)
-                                    | Value::Num(_)
-                                    | Value::Str(_)
-                                    | Value::Rat(_, _)
-                            )
-                            && body_entry_env.get_sym(*k) != Some(v)
-                        {
-                            captured_outer_writes.push(k.resolve().to_string());
-                        }
-                        merged.insert_sym(*k, v.clone());
-                    }
-                }
-            } else {
-                // Names bound from subsignature parameters are block-local and
-                // must not be written back to the caller (otherwise a parameter
-                // such as `|c(Str $x)` would clobber a caller variable that
-                // happens to share the parameter's name).
-                let mut subsig_names = std::collections::HashSet::new();
-                for pd in &data.param_defs {
-                    Self::collect_sub_signature_names(&pd.sub_signature, &mut subsig_names);
-                }
-                for (k, v) in self.env.iter() {
-                    if k == "_" || k == "@_" || subsig_names.contains(&k.resolve()) {
-                        continue;
-                    }
-                    if matches!(v, Value::Array(..)) {
-                        // Arrays are Arc-shared; in-place mutations are already
-                        // visible to the caller, and reassignment should propagate.
-                        merged.insert_sym(*k, v.clone());
-                    } else if merged.contains_key_sym(*k)
-                        && matches!(
-                            v,
-                            Value::Bool(_)
-                                | Value::Int(_)
-                                | Value::Num(_)
-                                | Value::Str(_)
-                                | Value::Rat(_, _)
-                        )
-                        // Only write a captured scalar back to the caller when the
-                        // body actually changed it from its body-entry value. A
-                        // value equal to the entry snapshot is a stale capture, not
-                        // a mutation, and must not clobber the caller's live value.
-                        && body_entry_env.get_sym(*k) != Some(v)
-                    {
-                        captured_outer_writes.push(k.resolve().to_string());
-                        merged.insert_sym(*k, v.clone());
-                    }
-                }
-            }
-            // Record the captured-outer scalar writes for precise caller-slot
-            // write-back (drained by the call-site `apply_pending_rw_writeback`).
-            self.pending_rw_writeback_sources
-                .extend(captured_outer_writes);
-            self.merge_sigilless_alias_writes(&mut merged, &self.env);
-            // Apply rw bindings after merge so they take precedence
-            self.apply_rw_bindings_to_env(&rw_bindings, &mut merged);
-            // Restore map rw topic tracker if it was set during block execution
-            if let Some(topic_val) = rw_map_topic {
-                merged.insert("__mutsu_rw_map_topic__".to_string(), topic_val);
-            }
-            self.env = merged;
-            self.restore_readonly_vars(saved_readonly);
-            if let Err(e) = &result
-                && e.is_fail()
-            {
-                return Ok(self.fail_error_to_failure_value(e));
-            }
-            // Bare blocks and pointy blocks are NOT routine boundaries for `return`.
-            // Propagate the return signal with the target callable ID so the
-            // enclosing routine can catch it.
-            let is_non_routine =
-                data.is_bare_block || data.compiled_code.as_ref().is_some_and(|cc| !cc.is_routine);
-            if is_non_routine
-                && let Err(ref e) = result
-                && e.return_value.is_some()
-            {
-                // Only propagate non-local return if we can identify the target
-                // routine (via __mutsu_callable_id in captured env or already set).
-                // If no target exists, catch it locally (e.g., supply block done+return).
-                let has_target = e.return_target_callable_id.is_some()
-                    || data.env.contains_key("__mutsu_callable_id");
-                if has_target {
-                    let mut e = result.unwrap_err();
-                    if e.return_target_callable_id.is_none()
-                        && let Some(Value::Int(id)) = data.env.get("__mutsu_callable_id")
-                    {
-                        e.return_target_callable_id = Some(*id as u64);
-                    }
-                    return Err(e);
-                }
-            }
-            let result = match result {
-                Err(e) if e.is_leave => return Err(e),
-                other => other,
-            };
-            let finalized =
-                self.finalize_return_with_spec(result, effective_return_spec.as_deref());
-            let fetch_rw = data.is_rw && !data.is_raw;
-            return finalized.and_then(|v| {
-                let v = if let Value::LazyList(list) = v {
-                    let mut env = list.env.clone();
-                    env.insert(
-                        "__mutsu_preserve_lazy_on_array_assign".to_string(),
-                        Value::Bool(true),
-                    );
-                    Value::LazyList(std::sync::Arc::new(crate::value::LazyList {
-                        body: list.body.clone(),
-                        env,
-                        cache: std::sync::Mutex::new(list.cache.lock().unwrap().clone()),
-                        compiled_code: list.compiled_code.clone(),
-                        compiled_fns: list.compiled_fns.clone(),
-                        elems_count: list.elems_count.clone(),
-                        scan_spec: list
-                            .scan_spec
-                            .as_ref()
-                            .map(|s| std::sync::Mutex::new(s.lock().unwrap().clone())),
-                        sequence_spec: list.sequence_spec.clone(),
-                        coroutine: list
-                            .coroutine
-                            .as_ref()
-                            .map(|c| std::sync::Mutex::new(c.lock().unwrap().clone())),
-                        lazy_pipe: list
-                            .lazy_pipe
-                            .as_ref()
-                            .map(|p| std::sync::Mutex::new(p.lock().unwrap().clone())),
-                        closure_seq: list
-                            .closure_seq
-                            .as_ref()
-                            .map(|c| std::sync::Mutex::new(c.lock().unwrap().clone())),
-                    }))
-                } else {
-                    v
-                };
-                self.maybe_fetch_rw_proxy(v, fetch_rw)
-            });
-        }
-        Err(RuntimeError::new("Callable expected"))
-    }
-
-    fn composed_result_to_args(value: Value, prefer_single: bool) -> Vec<Value> {
-        if prefer_single {
-            return match value {
-                Value::Seq(items) | Value::Slip(items) => vec![Value::array(items.to_vec())],
-                other => vec![other],
-            };
-        }
-        match value {
-            Value::Array(items, _) => items.to_vec(),
-            Value::Seq(items) => items.to_vec(),
-            Value::Slip(items) => items.to_vec(),
-            Value::Capture { positional, named } => {
-                let mut args = *positional;
-                for (k, v) in *named {
-                    args.push(Value::Pair(k, Box::new(v)));
-                }
-                args
-            }
-            other => vec![other],
-        }
-    }
-
-    /// Like `eval_block_value` but handles PRE/POST phasers in the body.
-    /// Used for function call evaluation where the body may contain PRE/POST.
-    pub(crate) fn eval_block_value_with_pre_post(
-        &mut self,
-        body: &[Stmt],
-    ) -> Result<Value, RuntimeError> {
-        let (pre_ph, _enter_ph, _success_ph, _failure_ph, post_ph, body_main) =
-            self.split_block_phasers(body);
-        if pre_ph.is_empty() && post_ph.is_empty() {
-            return self.eval_block_value(body);
-        }
-        // Run PRE phasers
-        for pre in &pre_ph {
-            let result = self.eval_block_value(std::slice::from_ref(pre))?;
-            if !result.truthy() {
-                return Err(Self::make_phaser_prepost_error(true));
-            }
-        }
-        // Run main body
-        let result = self.eval_block_value(&body_main);
-        // Run POST phasers (with $_ set to return value for POST checks)
-        let ret_val = match &result {
-            Ok(v) => v.clone(),
-            Err(e) => e.return_value.clone().unwrap_or(Value::Nil),
-        };
-        let saved_topic = self.env.get("_").cloned();
-        self.env.insert("_".to_string(), ret_val);
-        for post in &post_ph {
-            let post_result = self.eval_block_value(std::slice::from_ref(post));
-            match post_result {
-                Ok(v) if !v.truthy() => {
-                    if let Some(t) = saved_topic {
-                        self.env.insert("_".to_string(), t);
-                    }
-                    return Err(Self::make_phaser_prepost_error(false));
-                }
-                Err(e) => {
-                    if let Some(t) = saved_topic {
-                        self.env.insert("_".to_string(), t);
-                    }
-                    return Err(e);
-                }
-                _ => {}
-            }
-        }
-        if let Some(t) = saved_topic {
-            self.env.insert("_".to_string(), t);
-        }
-        result
-    }
-
-    /// Compile a block with `eval_block_value`'s compiler context (routine
-    /// scope, `$?PACKAGE`/`$?DISTRIBUTION`) without executing it. Pure
-    /// compilation — touches no `env`, runs no user code — so the VM can call it
-    /// (no env loan) and run the result in-place via `VM::run_nested` instead of
-    /// the `mem::take`/`VM::new` ping-pong (CP-3 collapse).
-    pub(crate) fn compile_block_value(
-        &self,
-        body: &[Stmt],
-    ) -> (
-        crate::opcode::CompiledCode,
-        std::collections::HashMap<String, crate::opcode::CompiledFunction>,
-    ) {
-        let mut compiler = crate::compiler::Compiler::new();
-        compiler.is_routine = !self.routine_stack.is_empty();
-        compiler.lexically_in_routine = !self.routine_stack.is_empty();
-        let scope = if let Some(frame) = self.routine_stack.last() {
-            // Set enclosing_package to the clean package name so that
-            // $?PACKAGE resolves to the package, not the mangled routine
-            // scope (e.g., "PackageTest" instead of "PackageTest::&pkg").
-            compiler.enclosing_package = Some(frame.package.clone());
-            format!("{}::&{}", frame.package, frame.name)
-        } else {
-            self.current_package()
-        };
-        compiler.set_current_package(scope);
-        // Resolve distribution context for $?DISTRIBUTION
-        compiler.current_distribution = self.current_distribution.clone().or_else(|| {
-            self.package_distributions
-                .get(&self.current_package())
-                .cloned()
-        });
-        compiler.compile(body)
-    }
-
-    pub(crate) fn eval_block_value(&mut self, body: &[Stmt]) -> Result<Value, RuntimeError> {
-        if body.is_empty() {
-            return Ok(Value::Nil);
-        }
-        let let_mark = self.let_saves_len();
-        let saved_functions = self.registry().functions.clone();
-        let saved_proto_subs = self.registry().proto_subs.clone();
-        let saved_proto_functions = self.registry().proto_functions.clone();
-        let saved_operator_assoc = self.operator_assoc.clone();
-        let saved_code_env: std::collections::HashMap<Symbol, Value> = self
-            .env
-            .iter()
-            .filter(|(k, _)| k.starts_with("&") || k.starts_with("__mutsu_callable_id::"))
-            .map(|(k, v)| (*k, v.clone()))
-            .collect();
-        let (code, compiled_fns) = self.compile_block_value(body);
-        // Multi-frame coherence (env_dirty-deletion path): box any captured-outer
-        // scalar this carrier body writes into a shared cell across env + saved
-        // frames, so the by-name write survives the owner frame's env restore.
-        // No-op in the default build (gated on cell_boxing_active).
-        self.box_carrier_free_var_writes(&code);
-        self.block_scope_depth += 1;
-        let result = self.run_compiled_block(&code, &compiled_fns);
-        let trailing_sub_value = match body.last() {
-            Some(Stmt::SubDecl {
-                name,
-                params,
-                param_defs,
-                body,
-                is_rw,
-                ..
-            }) => Some(Value::make_sub(
-                Symbol::intern(&self.current_package()),
-                *name,
-                params.clone(),
-                param_defs.clone(),
-                body.clone(),
-                *is_rw,
-                self.env.clone(),
-            )),
-            _ => None,
-        };
-        self.block_scope_depth = self.block_scope_depth.saturating_sub(1);
-        // Sub/proto declarations in block scope are lexical; restore registries on block exit.
-        self.registry_mut().functions = saved_functions;
-        self.registry_mut().proto_subs = saved_proto_subs;
-        self.registry_mut().proto_functions = saved_proto_functions;
-        self.operator_assoc = saved_operator_assoc;
-        self.env
-            .retain(|k, _| !(k.starts_with("&") || k.starts_with("__mutsu_callable_id::")));
-        for (k, v) in saved_code_env {
-            self.env.insert_sym(k, v);
-        }
-        // Blocks are scope boundaries for temp/let saves.
-        self.restore_let_saves(let_mark);
-        self.run_pending_instance_destroys()?;
-        result.map(|value| {
-            // When the block's last statement is a sub declaration, its value is
-            // the declared sub — not a leftover topic/`$_` value from an earlier
-            // statement (e.g. a Failure that propagated through a sink). Prefer
-            // the trailing sub unconditionally in that case.
-            if let Some(sub) = trailing_sub_value {
-                return sub;
-            }
-            value
-        })
-    }
-
-    /// Fast path for simple closures (e.g. sequence generators) that don't
-    /// declare subs or modify proto registries. Takes pre-compiled bytecode
-    /// to avoid recompilation on every call.
-    pub(crate) fn eval_precompiled_block_fast(
-        &mut self,
-        code: &crate::opcode::CompiledCode,
-        compiled_fns: &std::collections::HashMap<String, crate::opcode::CompiledFunction>,
-    ) -> Result<Value, RuntimeError> {
-        self.block_scope_depth += 1;
-        // CP-3 collapse: run the precompiled block re-entrantly in place.
-        let result = self.run_nested(code, compiled_fns);
-        for (slot, key) in &code.state_locals {
-            if let Some(name) = code.locals.get(*slot)
-                && let Some(val) = self.env().get(name).cloned()
-            {
-                self.set_state_var(key.clone(), val);
-            }
-        }
-        let value = self.env().get("_").cloned().unwrap_or(Value::Nil);
-        self.block_scope_depth = self.block_scope_depth.saturating_sub(1);
-        result.map(|last_value| last_value.unwrap_or(value))
-    }
-
-    /// Fast path for `Lock::Async.protect { ... }` — executes a bare block
-    /// directly in the current env without the full env save/restore overhead
-    /// of `call_sub_value`.  Shared vars must already be synced to env by the
-    /// caller.
-    pub(crate) fn call_protect_block(&mut self, code: &Value) -> Result<Value, RuntimeError> {
-        if let Value::Sub(data) = code {
-            let (compiled, compiled_fns, _captured_bindings, _writeback_bindings, captured_names) =
-                self.get_or_compile_protect_block_with_slots(data);
-            self.sync_shared_vars_for_names(captured_names.iter().map(|name| name.as_str()));
-            self.run_compiled_block(&compiled, compiled_fns.as_ref())
-        } else {
-            self.call_sub_value(code.clone(), Vec::new(), true)
-        }
-    }
-
-    pub(crate) fn get_or_compile_protect_block_with_slots(
-        &mut self,
-        data: &std::sync::Arc<crate::value::SubData>,
-    ) -> (
-        ProtectBlockCompiled,
-        ProtectBlockCompiledFns,
-        ProtectBlockCapturedBindings,
-        ProtectBlockWritebackBindings,
-        ProtectBlockCapturedNames,
-    ) {
-        let entry = self.protect_block_cache.entry(data.id).or_insert_with(|| {
-            let (compiled, compiled_fns) = if let Some(ref cc) = data.compiled_code {
-                (
-                    cc.clone(),
-                    std::sync::Arc::new(std::collections::HashMap::new()),
-                )
-            } else {
-                let compiler = crate::compiler::Compiler::new();
-                let (compiled, compiled_fns) = compiler.compile(&data.body);
-                (
-                    std::sync::Arc::new(compiled),
-                    std::sync::Arc::new(compiled_fns),
-                )
-            };
-            let captured_bindings: Vec<(usize, String)> = compiled
-                .locals
-                .iter()
-                .enumerate()
-                .filter(|(_, name)| data.env.contains_key(name))
-                .map(|(idx, name)| (idx, name.clone()))
-                .collect();
-            let mut assigned_slots = std::collections::HashSet::new();
-            for op in &compiled.ops {
-                match op {
-                    crate::opcode::OpCode::SetLocal(slot)
-                    | crate::opcode::OpCode::AssignExprLocal(slot)
-                    | crate::opcode::OpCode::PreIncrement(slot)
-                    | crate::opcode::OpCode::PreDecrement(slot)
-                    | crate::opcode::OpCode::PostIncrement(slot)
-                    | crate::opcode::OpCode::PostDecrement(slot) => {
-                        assigned_slots.insert(*slot as usize);
-                    }
-                    _ => {}
-                }
-            }
-            let writeback_bindings: Vec<(usize, String)> = captured_bindings
-                .iter()
-                .filter(|(slot, _)| assigned_slots.contains(slot))
-                .cloned()
-                .collect();
-            let mut captured_names: Vec<String> = captured_bindings
-                .iter()
-                .map(|(_, name)| name.clone())
-                .collect();
-            for op in &compiled.ops {
-                let name_idx = match op {
-                    crate::opcode::OpCode::GetGlobal(idx)
-                    | crate::opcode::OpCode::SetGlobal(idx)
-                    | crate::opcode::OpCode::GetArrayVar(idx)
-                    | crate::opcode::OpCode::GetHashVar(idx)
-                    | crate::opcode::OpCode::CheckReadOnly(idx) => Some(*idx as usize),
-                    _ => None,
-                };
-                let Some(idx) = name_idx else {
-                    continue;
-                };
-                let Some(crate::value::Value::Str(name)) = compiled.constants.get(idx) else {
-                    continue;
-                };
-                if data.env.contains_key(name.as_str()) && !captured_names.contains(name) {
-                    captured_names.push(name.to_string());
-                }
-            }
-            (
-                compiled,
-                compiled_fns,
-                std::sync::Arc::new(captured_bindings),
-                std::sync::Arc::new(writeback_bindings),
-                std::sync::Arc::new(captured_names),
-            )
-        });
-        (
-            entry.0.clone(),
-            entry.1.clone(),
-            entry.2.clone(),
-            entry.3.clone(),
-            entry.4.clone(),
-        )
-    }
-
-    /// Run pre-compiled bytecode and return the `$_` topic value.
-    fn run_compiled_block(
-        &mut self,
-        code: &crate::opcode::CompiledCode,
-        compiled_fns: &std::collections::HashMap<String, crate::opcode::CompiledFunction>,
-    ) -> Result<Value, RuntimeError> {
-        // CP-3 collapse: run the compiled block re-entrantly in place.
-        let result = self.run_nested(code, compiled_fns);
-        // Persist state variables for top-level compiled blocks too.
-        for (slot, key) in &code.state_locals {
-            if let Some(name) = code.locals.get(*slot)
-                && let Some(val) = self.env().get(name).cloned()
-            {
-                self.set_state_var(key.clone(), val);
-            }
-        }
-        let value = self.env().get("_").cloned().unwrap_or(Value::Nil);
-        result.map(|last_value| last_value.unwrap_or(value))
-    }
-
-    /// Slice F (env<->locals coherence): record an eager block-iteration's
-    /// captured-outer free-var writes so the call-site method op drains each new
-    /// value straight through to the caller's local slot, keeping it coherent
-    /// without the reverse `sync_locals_from_env` pull.
-    ///
-    /// The eager `.map`/`.grep` loops (`(1,2,3).map({ $sum += $_ })`) run the
-    /// block body via `run_reuse` rather than `call_compiled_closure_with_topic`,
-    /// so they bypass the closure-dispatch recording (#3307): a mutated captured
-    /// lexical lands in `env` but is never recorded for write-through. Replay the
-    /// block's compile-time `free_var_writes` here. Topic/param names are
-    /// loop-local (restored by the caller) and excluded.
-    pub(crate) fn record_eager_block_free_var_writeback(
-        &mut self,
-        code: &crate::opcode::CompiledCode,
-        params: &[String],
-    ) {
-        for sym in &code.free_var_writes {
-            sym.with_str(|fname| {
-                if fname != "_"
-                    && fname != "$_"
-                    && fname != "@_"
-                    && fname != "%_"
-                    && !params.iter().any(|p| p == fname)
-                {
-                    self.pending_rw_writeback_sources.push(fname.to_string());
-                }
-            });
-        }
-    }
-
-    pub(super) fn eval_map_over_items(
-        &mut self,
-        func: Option<Value>,
-        list_items: Vec<Value>,
-    ) -> Result<Value, RuntimeError> {
-        if let Some(Value::Sub(data)) = func {
-            let requires_full_binding = data.param_defs.iter().any(|pd| {
-                pd.named
-                    || pd.slurpy
-                    || pd.sigilless
-                    || pd.optional_marker
-                    || pd.default.is_some()
-                    || pd.type_constraint.is_some()
-                    || pd.where_constraint.is_some()
-                    || pd.sub_signature.is_some()
-                    || pd.outer_sub_signature.is_some()
-                    || pd.code_signature.is_some()
-                    || pd.shape_constraints.is_some()
-            });
-            if requires_full_binding {
-                let mut result = Vec::new();
-                for item in list_items {
-                    let value = self.call_sub_value(Value::Sub(data.clone()), vec![item], false)?;
-                    match value {
-                        Value::Slip(elems) => result.extend(elems.iter().cloned()),
-                        v => result.push(v),
-                    }
-                }
-                return Ok(Value::array(result));
-            }
-            let arity = if !data.params.is_empty() {
-                // Account for assumed positional args (from .assuming)
-                let effective = data
-                    .params
-                    .len()
-                    .saturating_sub(data.assumed_positional.len());
-                if effective == 0 { 1 } else { effective }
-            } else {
-                1
-            };
-            // Routine wrapper from .assuming() on a builtin — delegate to call_sub_value
-            // which knows how to resolve __mutsu_routine_name.
-            if data.env.contains_key("__mutsu_routine_name") {
-                let mut result = Vec::new();
-                let mut i = 0usize;
-                while i < list_items.len() {
-                    if arity > 1 && i + arity > list_items.len() {
-                        return Err(RuntimeError::new("Not enough elements for map block arity"));
-                    }
-                    let chunk: Vec<Value> = if arity == 1 {
-                        vec![list_items[i].clone()]
-                    } else {
-                        list_items[i..i + arity].to_vec()
-                    };
-                    let value = self.call_sub_value(Value::Sub(data.clone()), chunk, false)?;
-                    match value {
-                        Value::Slip(elems) => result.extend(elems.iter().cloned()),
-                        v => result.push(v),
-                    }
-                    i += arity;
-                }
-                return Ok(Value::array(result));
-            }
-            if data.env.contains_key("__mutsu_compose_left")
-                && data.env.contains_key("__mutsu_compose_right")
-            {
-                let mut result = Vec::new();
-                let mut i = 0usize;
-                while i < list_items.len() {
-                    if arity > 1 && i + arity > list_items.len() {
-                        return Err(RuntimeError::new("Not enough elements for map block arity"));
-                    }
-                    let chunk: Vec<Value> = if arity == 1 {
-                        vec![list_items[i].clone()]
-                    } else {
-                        list_items[i..i + arity].to_vec()
-                    };
-                    let value = self.call_sub_value(Value::Sub(data.clone()), chunk, false)?;
-                    match value {
-                        Value::Slip(elems) => result.extend(elems.iter().cloned()),
-                        v => result.push(v),
-                    }
-                    i += arity;
-                }
-                return Ok(Value::array(result));
-            }
-            let mut result = Vec::new();
-
-            // Extract LAST phaser bodies so they can be run after the map loop
-            let mut last_phaser_bodies: Vec<Vec<crate::ast::Stmt>> = Vec::new();
-            for stmt in &data.body {
-                if let crate::ast::Stmt::Phaser {
-                    kind: crate::ast::PhaserKind::Last,
-                    body,
-                } = stmt
-                {
-                    last_phaser_bodies.push(body.clone());
-                }
-            }
-
-            // Compile once, reuse VM for every iteration.
-            // `return` inside this block should propagate up to the
-            // lexically enclosing routine (if any), so mark the fresh
-            // compiler as being lexically nested inside a routine whenever
-            // a routine is currently on the dynamic call stack.
-            let mut compiler = crate::compiler::Compiler::new();
-            compiler.lexically_in_routine = !self.routine_stack.is_empty();
-            let (code, compiled_fns) = compiler.compile(&data.body);
-
-            let underscore = "_".to_string();
-            let dollar_topic = "$_".to_string();
-
-            // Save/restore only temporary bindings introduced by map itself.
-            // Captured lexical vars (in data.env) must keep mutations done inside
-            // the mapper block (e.g. `{ $a++ }`).
-            let mut touched_keys: Vec<String> = Vec::with_capacity(data.params.len() + 1);
-            for k in data.env.keys() {
-                if !self.env.contains_key_sym(*k) {
-                    touched_keys.push(k.resolve());
-                }
-            }
-            for p in &data.params {
-                if !touched_keys.contains(p) {
-                    touched_keys.push(p.clone());
-                }
-            }
-            if !touched_keys.iter().any(|k| k == "_") {
-                touched_keys.push(underscore.clone());
-            }
-            if !touched_keys.iter().any(|k| k == "$_") {
-                touched_keys.push(dollar_topic.clone());
-            }
-            let saved: Vec<(String, Option<Value>)> = touched_keys
-                .iter()
-                .map(|k| (k.clone(), self.env.get(k).cloned()))
-                .collect();
-
-            // Pre-insert closure env
-            for (k, v) in &data.env {
-                if !self.env.contains_key_sym(*k) {
-                    self.env.insert_sym(*k, v.clone());
-                }
-            }
-
-            // CP-3 collapse: run the map loop with fresh execution registers
-            // (replaces the `mem::take(self)` + `VM::new` sub-VM, reusing one
-            // register scope across all iterations). The closure returns the
-            // loop's Result; `with_nested_registers` restores the outer registers
-            // and flags env_dirty. The temporary-binding env restore (`saved`) is
-            // hoisted to after the call — it ran on every old exit path.
-            let loop_result: Result<Value, RuntimeError> = self.with_nested_registers(|vm| {
-                let mut i = 0usize;
-                while i < list_items.len() {
-                    if arity > 1 && i + arity > list_items.len() {
-                        return Err(RuntimeError::new("Not enough elements for map block arity"));
-                    }
-                    {
-                        let assumed_count = data.assumed_positional.len();
-                        // Bind assumed positional args first
-                        for (idx, val) in data.assumed_positional.iter().enumerate() {
-                            if let Some(p) = data.params.get(idx) {
-                                vm.env_mut().insert(p.clone(), val.clone());
-                            }
-                        }
-                        if arity == 1 {
-                            let item = list_items[i].clone();
-                            if let Some(p) = data.params.get(assumed_count) {
-                                vm.env_mut().insert(p.clone(), item.clone());
-                            }
-                            vm.env_mut().insert(underscore.clone(), item.clone());
-                            vm.env_mut().insert(dollar_topic.clone(), item);
-                        } else {
-                            for (idx, p) in data.params.iter().skip(assumed_count).enumerate() {
-                                if i + idx < list_items.len() {
-                                    vm.env_mut().insert(p.clone(), list_items[i + idx].clone());
-                                }
-                            }
-                            vm.env_mut()
-                                .insert(underscore.clone(), list_items[i].clone());
-                            vm.env_mut()
-                                .insert(dollar_topic.clone(), list_items[i].clone());
-                        }
-                    }
-                    match vm.run_reuse(&code, &compiled_fns) {
-                        Ok(()) => {
-                            let val = vm
-                                .last_stack_value()
-                                .cloned()
-                                .or_else(|| vm.env().get("_").cloned())
-                                .unwrap_or(Value::Nil);
-                            match val {
-                                Value::Slip(elems) => result.extend(elems.iter().cloned()),
-                                v => result.push(v),
-                            }
-                        }
-                        Err(e) if e.is_next() => {}
-                        Err(e) if e.is_last() => break,
-                        Err(mut e) => {
-                            // A `return` inside the map block targets the routine
-                            // that lexically encloses the block. Stamp the signal
-                            // with that routine's callable id (captured in the
-                            // closure env) so propagation/out-of-dynamic-scope
-                            // detection works when the map is forced lazily after
-                            // the enclosing routine has exited.
-                            if e.is_return()
-                                && e.return_target_callable_id.is_none()
-                                && let Some(Value::Int(id)) = data.env.get("__mutsu_callable_id")
-                            {
-                                e.return_target_callable_id = Some(*id as u64);
-                            }
-                            return Err(e);
-                        }
-                    }
-                    i += arity;
-                }
-
-                // Run LAST phasers after the map loop completes (natural end or `last`)
-                if !last_phaser_bodies.is_empty() && i > 0 {
-                    for phaser_body in &last_phaser_bodies {
-                        let phaser_compiler = crate::compiler::Compiler::new();
-                        let (phaser_code, phaser_fns) = phaser_compiler.compile(phaser_body);
-                        // Ignore errors from LAST phasers (best-effort)
-                        let _ = vm.run_reuse(&phaser_code, &phaser_fns);
-                        // Record the phaser's captured-outer writes (`LAST $x = True`)
-                        // so the enclosing call site drains them into the caller's
-                        // local slots — the map body's own writeback (below) only
-                        // covers `code`, not the separately-compiled phaser body.
-                        vm.record_eager_block_free_var_writeback(&phaser_code, &[]);
-                    }
-                }
-
-                Ok(Value::array(result))
-            });
-
-            // Restore original values (was done on every exit of the old loop).
-            for (k, orig) in saved {
-                match orig {
-                    Some(v) => {
-                        self.env.insert(k, v);
-                    }
-                    None => {
-                        self.env.remove(&k);
-                    }
-                }
-            }
-            if loop_result.is_ok() {
-                self.record_eager_block_free_var_writeback(&code, &data.params);
-            }
-            return loop_result;
-        }
-        if let Some(func) = func {
-            let mut result = Vec::new();
-            for item in list_items {
-                let value = self.call_sub_value(func.clone(), vec![item], false)?;
-                match value {
-                    Value::Slip(elems) => result.extend(elems.iter().cloned()),
-                    v => result.push(v),
-                }
-            }
-            return Ok(Value::array(result));
-        }
-        Ok(Value::array(list_items))
-    }
-
-    /// Like `eval_map_over_items` but writes back `$_` mutations to the source
-    /// list elements, implementing Raku's rw binding semantics for map.
-    /// Uses the same VM fast path as `eval_map_over_items` but checks for
-    /// `__mutsu_rw_map_topic__` after each iteration to capture mutations.
-    pub(super) fn eval_map_over_items_rw(
-        &mut self,
-        func: Option<Value>,
-        list_items: &mut [Value],
-    ) -> Result<Value, RuntimeError> {
-        let topic_key = "__mutsu_rw_map_topic__";
-        if let Some(Value::Sub(data)) = func {
-            let requires_full_binding = data.param_defs.iter().any(|pd| {
-                pd.named
-                    || pd.slurpy
-                    || pd.sigilless
-                    || pd.optional_marker
-                    || pd.default.is_some()
-                    || pd.type_constraint.is_some()
-                    || pd.where_constraint.is_some()
-                    || pd.sub_signature.is_some()
-                    || pd.outer_sub_signature.is_some()
-                    || pd.code_signature.is_some()
-                    || pd.shape_constraints.is_some()
-            });
-            if requires_full_binding
-                || data.env.contains_key("__mutsu_routine_name")
-                || data.env.contains_key("__mutsu_compose_left")
-            {
-                // Fall through to call_sub_value path for complex cases
-                let mut result = Vec::new();
-                let arity = if !data.params.is_empty() {
-                    let effective = data
-                        .params
-                        .len()
-                        .saturating_sub(data.assumed_positional.len());
-                    if effective == 0 { 1 } else { effective }
-                } else {
-                    1
-                };
-                let mut i = 0usize;
-                while i < list_items.len() {
-                    if arity > 1 && i + arity > list_items.len() {
-                        return Err(RuntimeError::new("Not enough elements for map block arity"));
-                    }
-                    let chunk: Vec<Value> = if arity == 1 {
-                        vec![list_items[i].clone()]
-                    } else {
-                        list_items[i..i + arity].to_vec()
-                    };
-                    self.env.remove(topic_key);
-                    let value = self.call_sub_value(Value::Sub(data.clone()), chunk, false)?;
-                    if arity == 1
-                        && let Some(mutated) = self.env.get(topic_key).cloned()
-                    {
-                        list_items[i] = mutated;
-                    }
-                    match value {
-                        Value::Slip(elems) => result.extend(elems.iter().cloned()),
-                        v => result.push(v),
-                    }
-                    i += arity;
-                }
-                self.env.remove(topic_key);
-                return Ok(Value::array(result));
-            }
-
-            let arity = if !data.params.is_empty() {
-                let effective = data
-                    .params
-                    .len()
-                    .saturating_sub(data.assumed_positional.len());
-                if effective == 0 { 1 } else { effective }
-            } else {
-                1
-            };
-            let mut result = Vec::new();
-
-            // Compile once, reuse VM for every iteration (same as eval_map_over_items)
-            let compiler = crate::compiler::Compiler::new();
-            let (code, compiled_fns) = compiler.compile(&data.body);
-
-            let underscore = "_".to_string();
-            let dollar_topic = "$_".to_string();
-
-            let mut touched_keys: Vec<String> = Vec::with_capacity(data.params.len() + 1);
-            for k in data.env.keys() {
-                if !self.env.contains_key_sym(*k) {
-                    touched_keys.push(k.resolve());
-                }
-            }
-            for p in &data.params {
-                if !touched_keys.contains(p) {
-                    touched_keys.push(p.clone());
-                }
-            }
-            if !touched_keys.iter().any(|k| k == "_") {
-                touched_keys.push(underscore.clone());
-            }
-            if !touched_keys.iter().any(|k| k == "$_") {
-                touched_keys.push(dollar_topic.clone());
-            }
-            let saved: Vec<(String, Option<Value>)> = touched_keys
-                .iter()
-                .map(|k| (k.clone(), self.env.get(k).cloned()))
-                .collect();
-
-            for (k, v) in &data.env {
-                if !self.env.contains_key_sym(*k) {
-                    self.env.insert_sym(*k, v.clone());
-                }
-            }
-
-            // CP-3 collapse: run the rw map loop with fresh execution registers
-            // (replaces the `mem::take(self)` + `VM::new` sub-VM). The closure
-            // returns the loop's Result; `with_nested_registers` restores the
-            // outer registers and flags env_dirty. The `saved`/`topic_key` env
-            // restore is hoisted to after the call (ran on every old exit).
-            let loop_result: Result<Value, RuntimeError> = self.with_nested_registers(|vm| {
-                let mut i = 0usize;
-                while i < list_items.len() {
-                    if arity > 1 && i + arity > list_items.len() {
-                        return Err(RuntimeError::new("Not enough elements for map block arity"));
-                    }
-                    {
-                        let assumed_count = data.assumed_positional.len();
-                        for (idx, val) in data.assumed_positional.iter().enumerate() {
-                            if let Some(p) = data.params.get(idx) {
-                                vm.env_mut().insert(p.clone(), val.clone());
-                            }
-                        }
-                        // Clear the topic tracker before each iteration
-                        vm.env_mut().remove(topic_key);
-                        if arity == 1 {
-                            let item = list_items[i].clone();
-                            if let Some(p) = data.params.get(assumed_count) {
-                                vm.env_mut().insert(p.clone(), item.clone());
-                            }
-                            vm.env_mut().insert(underscore.clone(), item.clone());
-                            vm.env_mut().insert(dollar_topic.clone(), item);
-                        } else {
-                            for (idx, p) in data.params.iter().skip(assumed_count).enumerate() {
-                                if i + idx < list_items.len() {
-                                    vm.env_mut().insert(p.clone(), list_items[i + idx].clone());
-                                }
-                            }
-                            vm.env_mut()
-                                .insert(underscore.clone(), list_items[i].clone());
-                            vm.env_mut()
-                                .insert(dollar_topic.clone(), list_items[i].clone());
-                        }
-                    }
-                    match vm.run_reuse(&code, &compiled_fns) {
-                        Ok(()) => {
-                            let val = vm
-                                .last_stack_value()
-                                .cloned()
-                                .or_else(|| vm.env().get("_").cloned())
-                                .unwrap_or(Value::Nil);
-                            // Write back topic mutation if it happened
-                            if arity == 1
-                                && let Some(mutated) = vm.env().get(topic_key).cloned()
-                            {
-                                list_items[i] = mutated;
-                            }
-                            match val {
-                                Value::Slip(elems) => result.extend(elems.iter().cloned()),
-                                v => result.push(v),
-                            }
-                        }
-                        Err(e) if e.is_next() => {
-                            if arity == 1
-                                && let Some(mutated) = vm.env().get(topic_key).cloned()
-                            {
-                                list_items[i] = mutated;
-                            }
-                        }
-                        Err(e) if e.is_last() => {
-                            if arity == 1
-                                && let Some(mutated) = vm.env().get(topic_key).cloned()
-                            {
-                                list_items[i] = mutated;
-                            }
-                            break;
-                        }
-                        Err(e) => {
-                            return Err(e);
-                        }
-                    }
-                    i += arity;
-                }
-
-                Ok(Value::array(result))
-            });
-
-            for (k, orig) in saved {
-                match orig {
-                    Some(v) => self.env.insert(k, v),
-                    None => self.env.remove(&k),
-                };
-            }
-            self.env.remove(topic_key);
-            return loop_result;
-        }
-        // Non-Sub func: delegate to regular map
-        self.eval_map_over_items(func, list_items.to_vec())
-    }
-
-    pub(super) fn eval_grep_over_items_with_mutated(
-        &mut self,
-        func: Option<Value>,
-        mut list_items: Vec<Value>,
-    ) -> Result<(Value, Vec<Value>), RuntimeError> {
-        if let Some(Value::Sub(data)) = func {
-            let mut result = Vec::new();
-            let arity = if !data.params.is_empty() {
-                let effective = data
-                    .params
-                    .len()
-                    .saturating_sub(data.assumed_positional.len());
-                if effective == 0 { 1 } else { effective }
-            } else {
-                1
-            };
-            if data.env.contains_key("__mutsu_compose_left")
-                && data.env.contains_key("__mutsu_compose_right")
-            {
-                let mut i = 0usize;
-                while i < list_items.len() {
-                    if arity > 1 && i + arity > list_items.len() {
-                        break;
-                    }
-                    let chunk: Vec<Value> = if arity == 1 {
-                        vec![list_items[i].clone()]
-                    } else {
-                        list_items[i..i + arity].to_vec()
-                    };
-                    let pred =
-                        self.call_sub_value(Value::Sub(data.clone()), chunk.clone(), false)?;
-                    if pred.truthy() {
-                        if arity == 1 {
-                            result.push(chunk[0].clone());
-                        } else {
-                            result.push(Value::array(chunk));
-                        }
-                    }
-                    i += arity;
-                }
-                return Ok((Value::array(result), list_items));
-            }
-
-            // Compile once, reuse VM for every iteration.
-            // `return` inside this block should propagate up to the
-            // lexically enclosing routine (if any), so mark the fresh
-            // compiler as being lexically nested inside a routine whenever
-            // a routine is currently on the dynamic call stack.
-            let mut compiler = crate::compiler::Compiler::new();
-            compiler.lexically_in_routine = !self.routine_stack.is_empty();
-            let (code, compiled_fns) = compiler.compile(&data.body);
-
-            let underscore = "_".to_string();
-            let dollar_topic = "$_".to_string();
-
-            let mut touched_keys: Vec<String> = Vec::with_capacity(data.params.len() + 2);
-            for k in data.env.keys() {
-                if !self.env.contains_key_sym(*k) {
-                    touched_keys.push(k.resolve());
-                }
-            }
-            for p in &data.params {
-                if !touched_keys.contains(p) {
-                    touched_keys.push(p.clone());
-                }
-            }
-            if !touched_keys.iter().any(|k| k == "_") {
-                touched_keys.push(underscore.clone());
-            }
-            if !touched_keys.iter().any(|k| k == "$_") {
-                touched_keys.push(dollar_topic.clone());
-            }
-            let topic_source_key = "__mutsu_grep_topic_source".to_string();
-            if !touched_keys.iter().any(|k| k == &topic_source_key) {
-                touched_keys.push(topic_source_key.clone());
-            }
-            let saved: Vec<(String, Option<Value>)> = touched_keys
-                .iter()
-                .map(|k| (k.clone(), self.env.get(k).cloned()))
-                .collect();
-
-            // Pre-insert closure env
-            for (k, v) in &data.env {
-                self.env.insert_sym(*k, v.clone());
-            }
-
-            // CP-3 collapse: run the grep loop with fresh execution registers
-            // (replaces the `mem::take(self)` + `VM::new` sub-VM). The closure
-            // returns Ok(()) / Err on the loop; `with_nested_registers` restores
-            // the outer registers and flags env_dirty. The `saved` env restore is
-            // hoisted to after the call (ran on every old exit path).
-            let loop_result: Result<(), RuntimeError> = self.with_nested_registers(|vm| {
-                let mut i = 0usize;
-                let mut stop = false;
-                while i < list_items.len() {
-                    if arity > 1 && i + arity > list_items.len() {
-                        break;
-                    }
-                    let chunk: Vec<Value> = if arity == 1 {
-                        vec![list_items[i].clone()]
-                    } else {
-                        list_items[i..i + arity].to_vec()
-                    };
-                    'body_redo: loop {
-                        {
-                            let assumed_count = data.assumed_positional.len();
-                            for (idx, val) in data.assumed_positional.iter().enumerate() {
-                                if let Some(p) = data.params.get(idx) {
-                                    vm.env_mut().insert(p.clone(), val.clone());
-                                }
-                            }
-                            if arity == 1 {
-                                if let Some(p) = data.params.get(assumed_count) {
-                                    vm.env_mut().insert(p.clone(), chunk[0].clone());
-                                }
-                                vm.env_mut().insert(underscore.clone(), chunk[0].clone());
-                                vm.env_mut().insert(dollar_topic.clone(), chunk[0].clone());
-                                vm.env_mut()
-                                    .insert(topic_source_key.clone(), chunk[0].clone());
-                            } else {
-                                for (idx, p) in data.params.iter().skip(assumed_count).enumerate() {
-                                    if idx < chunk.len() {
-                                        vm.env_mut().insert(p.clone(), chunk[idx].clone());
-                                    }
-                                }
-                                vm.env_mut().insert(underscore.clone(), chunk[0].clone());
-                                vm.env_mut().insert(dollar_topic.clone(), chunk[0].clone());
-                            }
-                        }
-                        vm.set_topic_source_var((arity == 1).then_some(topic_source_key.clone()));
-                        match vm.run_reuse(&code, &compiled_fns) {
-                            Ok(()) => {
-                                let pred = vm
-                                    .last_stack_value()
-                                    .cloned()
-                                    .or_else(|| vm.env().get("_").cloned())
-                                    .unwrap_or(Value::Nil);
-                                let updated_item = if arity == 1 {
-                                    vm.env()
-                                        .get(&topic_source_key)
-                                        .cloned()
-                                        .unwrap_or_else(|| chunk[0].clone())
-                                } else {
-                                    chunk[0].clone()
-                                };
-                                if arity == 1 {
-                                    list_items[i] = updated_item.clone();
-                                }
-                                if pred.truthy() {
-                                    if arity == 1 {
-                                        result.push(updated_item);
-                                    } else {
-                                        result.push(Value::array(chunk));
-                                    }
-                                }
-                                break 'body_redo;
-                            }
-                            Err(e) if e.is_redo() => continue 'body_redo,
-                            Err(e) if e.is_next() => break 'body_redo,
-                            Err(e) if e.is_last() => {
-                                stop = true;
-                                break 'body_redo;
-                            }
-                            Err(e) => {
-                                return Err(e);
-                            }
-                        }
-                    }
-                    if stop {
-                        break;
-                    }
-                    i += arity;
-                }
-                Ok(())
-            });
-
-            for (k, orig) in saved {
-                match orig {
-                    Some(v) => {
-                        self.env.insert(k, v);
-                    }
-                    None => {
-                        self.env.remove(&k);
-                    }
-                }
-            }
-            if loop_result.is_ok() {
-                self.record_eager_block_free_var_writeback(&code, &data.params);
-            }
-            loop_result?;
-            return Ok((Value::array(result), list_items));
-        }
-        if let Some(pattern) = func {
-            if matches!(pattern, Value::Bool(_)) {
-                return Err(RuntimeError::new("X::Match::Bool"));
-            }
-            let mut result = Vec::new();
-            for item in &list_items {
-                if self.smart_match(item, &pattern) {
-                    result.push(item.clone());
-                }
-            }
-            return Ok((Value::array(result), list_items));
-        }
-        if let Some(func) = func {
-            let mut result = Vec::new();
-            for item in &list_items {
-                let pred = self.call_sub_value(func.clone(), vec![item.clone()], false)?;
-                if pred.truthy() {
-                    result.push(item.clone());
-                }
-            }
-            return Ok((Value::array(result), list_items));
-        }
-        Ok((Value::array(list_items.clone()), list_items))
-    }
-
-    pub(super) fn eval_grep_over_items(
-        &mut self,
-        func: Option<Value>,
-        list_items: Vec<Value>,
-    ) -> Result<Value, RuntimeError> {
-        let (result, _) = self.eval_grep_over_items_with_mutated(func, list_items)?;
-        Ok(result)
-    }
-
-    pub(super) fn find_first_match_over_items(
-        &mut self,
-        func: Option<Value>,
-        list_items: &[Value],
-        from_end: bool,
-    ) -> Result<Option<(usize, Value)>, RuntimeError> {
-        let mut matcher = InterpFirstMatcher(self);
-        find_first_match_generic(&mut matcher, func.as_ref(), list_items, from_end)
-    }
-}
-
-/// Engine-agnostic matcher for `.first` / `first`: decides whether a single
-/// element matches the pattern. The only engine-specific part is invoking a
-/// `Sub` matcher block (the genuine Category-B fork); `smart_match` for non-block
-/// patterns is the interpreter's single shared implementation. Implemented for
-/// both the interpreter ([`InterpFirstMatcher`]) and the VM (`VmFirstMatcher`),
-/// so the iteration / `:end` / `pair_as_positional` logic lives once in
-/// [`find_first_match_generic`].
-pub(crate) trait FirstMatcher {
-    fn item_matches(&mut self, pattern: &Value, item: &Value) -> Result<bool, RuntimeError>;
-}
-
-/// [`FirstMatcher`] backed by the tree-walking interpreter.
-pub(crate) struct InterpFirstMatcher<'a>(pub(crate) &'a mut Interpreter);
-
-impl FirstMatcher for InterpFirstMatcher<'_> {
-    fn item_matches(&mut self, pattern: &Value, item: &Value) -> Result<bool, RuntimeError> {
-        if matches!(pattern, Value::Sub(_)) {
-            // Pass the element positionally: a Hash-sourced `Value::Pair` would
-            // otherwise bind as a named arg, leaving the block with zero
-            // positionals (`%h.first({ .value > 1 })`).
-            let call_item = crate::runtime::utils::pair_as_positional(item);
-            Ok(self
-                .0
-                .call_sub_value(pattern.clone(), vec![call_item], true)?
-                .truthy())
-        } else {
-            Ok(self.0.smart_match(item, pattern))
-        }
-    }
-}
-
-/// Shared `.first` / `first` scan: return the first (or, with `from_end`, last)
-/// `(index, value)` whose element matches `pattern` (or any element when
-/// `pattern` is `None`). Engines supply matching through `matcher`.
-pub(crate) fn find_first_match_generic(
-    matcher: &mut dyn FirstMatcher,
-    pattern: Option<&Value>,
-    list_items: &[Value],
-    from_end: bool,
-) -> Result<Option<(usize, Value)>, RuntimeError> {
-    if list_items.is_empty() {
-        return Ok(None);
-    }
-    let len = list_items.len();
-    for idx in 0..len {
-        let actual_idx = if from_end { len - 1 - idx } else { idx };
-        let item = list_items[actual_idx].clone();
-        let matched = match pattern {
-            Some(p) => matcher.item_matches(p, &item)?,
-            None => true,
-        };
-        if matched {
-            return Ok(Some((actual_idx, item)));
-        }
-    }
-    Ok(None)
 }
